Major refactoring of IO-handling. Refactored to use Externalizable, added LoadingScreen + OSMHandler memory optimizations

+++ src/model/osm/OSMWay.java
@@ -3,10 +3,8 @@ package model.osm;
 import java.util.ArrayList;
 
 public class OSMWay extends ArrayList<OSMNode>{
-    private double xCoord;
-    private double yCoord;
-
-
+    private float xCoord;
+    private float yCoord;
 
     public OSMWay() {
     }
